refactor(harness): split middleware/library/mod.rs by concern
library/mod.rs mixed 14 built-in middleware across 4 unrelated
concerns in ~1450 lines. Split mechanically into resilience.rs
(retry/timeout/fallback/rate-limit), budget.rs (token/cost tracking
and enforcement), tool_policy.rs (allowlisting, policy, dynamic/
contextual selection, human approval), and observe.rs (structured-
output validation, dynamic prompt, redaction, tracing). No behavior
change.
